child,testfiles,pids: fix %u/%d printf format mismatches
Three snprintf call sites used %u to print signed integer
expressions: rand() % 8, 1 + (rand() % MAX_TESTFILES), and a
pid_t pid value. The format/argument signedness mismatch is
undefined behavior in principle and produces wrong output for
negative values. Switch all three to %d, casting pid to int
to match the conversion explicitly.

+++ pids.c
@@ -163,7 +163,7 @@ void dump_childnos(void)
 			} else {
 				pid_t pid = __atomic_load_n(&pids[i + j], __ATOMIC_RELAXED);
 
-				n = snprintf(sptr, end - sptr, "%u ", pid);
+				n = snprintf(sptr, end - sptr, "%d ", (int)pid);
 			}
 			if (n > 0 && n < end - sptr)
 				sptr += n;
